feat(website): route /app to the web app, stop promoting native download
Replace the /app TestFlight/APK page with a server-side redirect to
https://app.meetwithoutfear.com/ so visitors land directly in the hosted
web experience. Move the old download UI to /download, unlinked from the
marketing surface, so the in-app banner's fallback has somewhere to go.
Rename home CTA copy from "Download the App" to "Get Started" and update
the invitation SuccessState's secondary link to point at /download.
